pyproject.toml
[project]
name = "worddocautomation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx2pdf>=0.1.8",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-docx>=1.2.0",
]
